requirements.txt
python-hglib>=2.4
whatthepatch>=0.0.4
